Conflicts:
	src/main/java/com/clevel/selos/businesscontrol/BankStmtControl.java
	src/main/java/com/clevel/selos/controller/BizInfoSummary.java
	src/main/java/com/clevel/selos/controller/WelcomePage.java
	src/main/java/com/clevel/selos/integration/dwh/DWHInterfaceImpl.java
	src/main/java/com/clevel/selos/integration/ncb/ncrs/service/NCRSImp.java
	src/main/java/com/clevel/selos/model/db/ext/dwh/Obligation1.java
	src/main/java/com/clevel/selos/model/view/CustomerInfoSummaryView.java

+++ src/main/java/com/clevel/selos/dao/working/PrescreenDAO.java
@@ -9,7 +9,7 @@ import org.slf4j.Logger;
 
 import javax.inject.Inject;
 
-public class PrescreenDAO extends GenericDAO<Prescreen, Integer> {
+public class PrescreenDAO extends GenericDAO<Prescreen, Long> {
     @Inject
     private Logger log;
